Revert "Differenciate between `MutationGuard` implementation requirements"
This reverts commit a45536ef548fd8c73e485b01dff3ea1014dd7bb7.
